requirements.txt
autogen-agentchat>=0.2.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
openseespy>=3.6.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
pint>=0.23
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
tiktoken>=0.5.0
questionary>=2.0.0
rich>=13.0.0
typer>=0.9.0
tqdm>=4.66.0 